removed the _to_xml function from automated-tasks.py to json_to_xml.py
